Extract Search lazy components into colocated module
Search.tsx had 75 lines of lazy declarations and skeleton fallbacks
that aren't part of the page's orchestration. Move them to
client/pages/search/lazy.tsx so the route shell stays focused.

Search.tsx down from 1,381 to 1,295 lines. The meaty subcomponent
split (header, active filters bar, per-mode results panels) and the
imperative-to-declarative conversion of the three search hooks
remain — those are coupled to a UX decision (search-on-submit vs
search-as-you-type) that deserves a dedicated follow-up.
